+++ index.html
@@ -164,6 +164,6 @@
       </div>
     </section>
   </template>
-<script src="js/main.js"></script>
+<script type="module" src="js/main.js"></script>
 </body>
 </html>
